03_web_app/requirements.txt
# Core Framework
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1

# Database
sqlalchemy==2.0.28

# API Integrations
openai==1.12.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0

# Utilities
python-dotenv==1.0.1
pydantic==1.10.13
python-multipart==0.0.9

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5 